Add Appveyor build configuration ...
Add Appveyor build configuration to enable building and testing the
extension on Windows.

Note: The current configuration is fetching a pull request for
bitcoin/libbase58 that resolves issues when building with Visual Studio.

Note: The tests are currently being skipped. Some investigation will be
required to determine why the tests are skipped.